+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `EncodeLabelSet` is now implemented for tuples `(A: EncodeLabelSet, B: EncodeLabelSet)`.
    See [PR 257].
 
-- `Family::get_or_create_clone` can access a metric in a labeled family. This
+- `Family::get_or_create_owned` can access a metric in a labeled family. This
    method avoids the risk of runtime deadlocks at the expense of creating an
    owned type. See [PR 244].
+ 
+ [PR 244]: https://github.com/prometheus/client_rust/pull/244
 
 ### Changed